refactor: yesterday 메서드로 통합
