+++ BoxOffice/App/SceneDelegate.swift
@@ -20,7 +20,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         let navigationController = UINavigationController(rootViewController: mainViewController)
         window?.rootViewController = navigationController
         navigationController.view.backgroundColor = .white
-//        window?.backgroundColor = .white
         window?.makeKeyAndVisible()
     }
